Add appconfig back with appropriate yaml config
The distributed calculator quickstart is referenced as an example
in the observability quickstart. This commit adds the appconfig
back into the apps configuration set and introduces the actual
appconfig as well to allow the app to start. This lets users
experience the simplest path for the observability quickstart
while unblocking the calculator.